Practical Byzantine Fault Tolerance PBFT
Practical Byzantine Fault Tolerance (PBFT) consensus algorithm.
we need to make several changes. PBFT is a consensus algorithm that allows a distributed system to reach agreement even in the presence of faulty or malicious nodes. 
It works by having a leader (primary) node propose a block, and the other nodes (replicas) vote on the proposal. 
If a sufficient number of nodes agree, the block is added to the blockchain.

Explanation of Changes:
Validators and Primary: The validators list contains the nodes participating in the consensus. The primary is the leader node that proposes the block.

PBFT Consensus: The pbft_consensus method implements the PBFT algorithm:

Pre-Prepare: The primary proposes a block.

Prepare: Validators vote on the proposed block.

Commit: If enough validators agree, they commit the block.

Fault Tolerance: The f variable represents the maximum number of faulty nodes. For PBFT to work, at least 2f + 1 validators must agree on the block.

Voting Simulation: Validators vote with a 90% chance of agreeing (simulating honest nodes).